Move zipper implementation to its own file and document it
